add onboard screen
