+++ lib/main.dart
@@ -1,36 +1,46 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+import 'package:termomete/screens/onboarding_screen.dart';
 import 'package:termomete/widget/navigation.dart';
 
-import 'about.dart';
-import 'addfreezer.dart';
-import 'createprofile.dart';
-import 'editfreezerdetails.dart';
-import 'freezerdetails.dart';
-import 'help.dart';
-import 'home.dart';
-import 'intro_search.dart';
-import 'language.dart';
-import 'login.dart';
-import 'newpassword.dart';
-import 'notification.dart';
-import 'profile.dart';
-import 'search.dart';
-import 'setting.dart';
-import 'signup.dart';
-import 'cpassword.dart';
-import 'splash.dart';
-import 'successchange.dart';
-import 'verifyemail.dart';
+import 'screens/about.dart';
+import 'screens/addfreezer.dart';
+import 'screens/createprofile.dart';
+import 'screens/editfreezerdetails.dart';
+import 'screens/freezerdetails.dart';
+import 'screens/help.dart';
+import 'screens/home.dart';
+import 'screens/intro_search.dart';
+import 'screens/language.dart';
+import 'screens/login.dart';
+import 'screens/newpassword.dart';
+import 'screens/notification.dart';
+import 'screens/profile.dart';
+import 'screens/search.dart';
+import 'screens/setting.dart';
+import 'screens/signup.dart';
+import 'screens/cpassword.dart';
+import 'screens/splash.dart';
+import 'screens/successchange.dart';
+import 'screens/verifyemail.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
-  runApp(MyApp());
+  final pref = await SharedPreferences.getInstance();
+  final showHome = pref.getBool('showHome') ?? false;
+  runApp(MyApp(
+    showHome: showHome,
+  ));
 }
 
 class MyApp extends StatefulWidget {
+  bool showHome;
+  MyApp({required this.showHome});
+
   @override
   State<MyApp> createState() => _MyAppState();
 }
@@ -69,12 +79,12 @@ class _MyAppState extends State<MyApp> {
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
-        home: CurrentPAge());
+        home: widget.showHome ? CurrentPage() : OnboardingPage());
   }
 }
 
-class CurrentPAge extends StatelessWidget {
-  const CurrentPAge({super.key});
+class CurrentPage extends StatelessWidget {
+  const CurrentPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -87,7 +97,7 @@ class CurrentPAge extends StatelessWidget {
           } else if (snapshot.connectionState == ConnectionState.waiting) {
             return CircularProgressIndicator();
           } else {
-            return AddFreezerPage();
+            return HomePage();
           }
         },
       ),
