+++ lib/screens/editfreezerdetails.dart
@@ -1,22 +1,19 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 
-import 'widget/Custom_button.dart';
-import 'widget/Custom_text.dart';
-import 'widget/Cutom_inputField.dart';
+import '/widget/Custom_button.dart';
+import '/widget/Custom_text.dart';
+import '/widget/Cutom_inputField.dart';
 
-class AddFreezerPage extends StatefulWidget {
-  const AddFreezerPage({Key? key}) : super(key: key);
+class EditFreezerPage extends StatefulWidget {
+  const EditFreezerPage({Key? key}) : super(key: key);
 
   @override
-  State<AddFreezerPage> createState() => _AddFreezerPageState();
+  State<EditFreezerPage> createState() => _EditFreezerPageState();
 }
 
-class _AddFreezerPageState extends State<AddFreezerPage> {
+class _EditFreezerPageState extends State<EditFreezerPage> {
   TextEditingController _freezerName = TextEditingController();
   TextEditingController _items = TextEditingController();
-  final formKey = GlobalKey<FormState>();
   @override
   void dispose() {
     _freezerName.dispose();
@@ -25,6 +22,7 @@ class _AddFreezerPageState extends State<AddFreezerPage> {
     super.dispose();
   }
 
+  final formKey = GlobalKey<FormState>();
   String dropdownvalue = 'Option 1';
   List<String> items = ['Option 1', 'Option 2', 'Option 3', 'Option 4'];
   @override
@@ -35,10 +33,12 @@ class _AddFreezerPageState extends State<AddFreezerPage> {
       backgroundColor: Color.fromRGBO(244, 246, 254, 1),
       appBar: AppBar(
         backgroundColor: Color.fromRGBO(11, 55, 120, 1),
-        leading: backButton(path: '/bnavigation'),
+        leading: backButton(
+          path: '/home',
+        ),
         centerTitle: true,
         title: CustomText(
-          text: 'Add New Freezer',
+          text: 'Edit Freezer Details',
         ),
         actions: [
           closebutton(),
@@ -58,8 +58,7 @@ class _AddFreezerPageState extends State<AddFreezerPage> {
                 height: h * 0.01,
               ),
               inputField(
-                validator: (name) =>
-                    name!.isEmpty ? 'enter freezer name' : null,
+                validator: (name) => name == null ? 'enter freezer name' : null,
                 controller: _freezerName,
                 hintText: '',
               ),
@@ -102,26 +101,22 @@ class _AddFreezerPageState extends State<AddFreezerPage> {
                 height: h * 0.01,
               ),
               inputField(
-                validator: (items) {
-                  if (items!.isEmpty) {
-                    return 'enter items';
-                  } else
-                    return null;
-                },
+                validator: (name) => name == null ? 'enter items' : null,
                 controller: _items,
                 hintText: '',
-                maxLines: 3,
+                maxLines: 2,
               ),
               SizedBox(
-                height: h * 0.05,
+                height: h * 0.07,
               ),
               CustomSquareButton(
                 onTap: () {
                   final isValid = formKey.currentState!.validate();
                   if (!isValid) return;
-                  Navigator.of(context).pushNamed('/home');
+
+                  Navigator.of(context).pushNamed('/cprofile');
                 },
-                buttonText: 'Save',
+                buttonText: 'Save Edited Details',
               ),
             ],
           ),
