+++ pubspec.yaml
@@ -31,6 +31,14 @@ dependencies:
   firebase_core: ^2.15.1
   firebase_auth: ^4.9.0
   email_validator: ^2.1.17
+  firebase_database: ^10.2.5
+  google_sign_in: ^6.1.5
+
+  # page indicator
+  smooth_page_indicator: ^1.0.0+2
+
+  # storage to show onboarding one time only
+  shared_preferences: ^2.0.12
 
 dev_dependencies:
   flutter_test:
